feature/JKP-0008 - Criando um package e modulos para organização de código
